reviewai_railway_saas/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.29.0
jinja2>=3.1.4
sqlalchemy>=2.0.25
pydantic>=2.5.3
pydantic-settings>=2.0.3
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
requests>=2.31.0
apscheduler>=3.10.4
reportlab>=4.0.9
matplotlib>=3.8.2
pandas>=2.2.0
PyPDF2>=3.0.1